Brand link adjusted in navbar

+++ db/migrate/20170907170059_create_items.rb
@@ -2,7 +2,7 @@ class CreateItems < ActiveRecord::Migration[5.0]
   def change
     create_table :items do |t|
       t.string :name
-      t.integer :quantity
+      t.decimal :quantity
       t.string :quantity_type
       t.references :user, foreign_key: true
